requirements.txt
south
simplemysql
django<=1.6
django-grappelli
django-filebrowser
django-registration
MYSQL-python
django-jenkins
